Restructure modules into proper folder
