+++ lib/immutable_struct_ex/comparable.rb
