+++ lib/immutable_struct_ex/immutable.rb
